separated backend and python interface to backend
