+++ pintool/makefile.rules
@@ -16,7 +16,7 @@
 
 CACHESIMDIR := ../cachesim/
 
-TOOL_CXXFLAGS += -DNO_PYTHON -I../cachesim/ 
+TOOL_CXXFLAGS += -DNO_PYTHON -I$(CACHESIMDIR)
 
 #PYTHON_INCLUDE_PATH := /usr/include/python2.7
 PYTHON_INCLUDE_PATH := /usr/include/python3.6m/
@@ -83,7 +83,7 @@ SANITY_SUBSET := $(TEST_TOOL_ROOTS) $(TEST_ROOTS)
 $(OBJDIR)cachesim_itf$(PINTOOL_SUFFIX): $(OBJDIR)cachesim_itf$(OBJ_SUFFIX) $(OBJDIR)backend$(OBJ_SUFFIX)
 	$(LINKER) $(TOOL_LDFLAGS) $(LINK_EXE)$@ $^ $(TOOL_LPATHS) $(TOOL_LIBS)
 
-$(OBJDIR)backend$(OBJ_SUFFIX): $(CACHESIMDIR)backend.c
-	$(APP_CC) -DNDEBUG -g -O3 -Wall -Wstrict-prototypes -fPIC -I$(PYTHON_INCLUDE_PATH) -DNO_PYTHON -c $(CACHESIMDIR)backend.c -o $(OBJDIR)backend$(OBJ_SUFFIX)
+$(OBJDIR)backend$(OBJ_SUFFIX): $(CACHESIMDIR)backend.c $(CACHESIMDIR)backend.h
+	$(APP_CC) -DNDEBUG -g -O3 -Wall -Wstrict-prototypes -fPIC -DNO_PYTHON -c $(CACHESIMDIR)backend.c -o $(OBJDIR)backend$(OBJ_SUFFIX)
 
 # gcc -DNDEBUG -g -O3 -Wall -Wstrict-prototypes -fPIC -I/usr/include/python2.7 -c backend.c -o cachesim.o
